split out platform thread implementation
